requirements.txt
flask~=2.3.2
git+https://github.com/DAMcraft/async_dns
mysql-connector-python
python-dotenv
requests
